Update and testing of AlignmentDistanceCalculator
Fixed up parts of test_Base and added a tearDownClass. Tested AlignmentDistanceCalculator, had to make correction for how gaps were being handled which was throwing off the pairwise distance calculation involving scoring matrices. Had to also correct the ETMIPWrapper because it relied on a method to convert a np.array to a DistanceMatrix which used to be part of SeqAlignment. I found it in the git history and added it back into AlignmentDistanceCalculator as a static method.