Corrected PhylogeneticTree behavior and testing
For traversals I corrected the way that the branch path length was calculated in the class and in the tests. For the bottom up traversal I also corrected how the queue operates, passing not only the node of interest and its path length but also its actual path. This allowed me to update how parents are added and removed to the queue which led to a correct traversal. For testing tree agreement I have used a some what relaxed requirement because it is possible for leaf nodes at the same level to switch places. To test tree agreement I now group all nodes at the same path length together and sort them. I then require the sorted nodes to agree. This means the exact order in the tree may be slightly different but the structure will be the same.