Working on PhylogeneticTree
Added a new attribute size. This is set when a tree is constructed by getting the size of the distance matrix, it is not expected that a distance matrix be passed in for custom trees, or that a list of terminals (like SeqAlignment.seq_order) be provided so that the right size can be measured. Added additional methods: rename_internal_nodes, assign_group_rank, and traverse_by_rank. rename_internal_nodes is used to consistently name all internal nodes in the produced tree no matter which construction method is used. Assign_group_rank is used to generate a dictionary mapping the rank (level in the tree) and group (cluster within a rank) to a root node and a list of associated terminals (sequence identifiers). traverse_by_rank traverses the tree from root to leaves returning a list of nodes at each level which has the same length as the rank (1=root, n=leaves). The traversal is achieved by always replacing the shallowest internal node with its two children. Still in the process of testing these methods, having some issues getting assign_group_rank to match up with the ETC version.