Working on PhylogeneticTree
Updated the custom tree construction (import nhx tree file) method. It now only renames if there is a confidence value for the node. I should also check if there is already a name or not. I also updated the rename nodes method (Still needs to be tested). The biggest changes are in the traverse_by_rank method. This now traverses by node name instead of trying to use branch lengths or node depths. This assumes that internal nodes are named in ascending order (ETC default) or descending order (biopython default).