Working on DataSetGenerator
Updated how PDB sequences are parsed, by using the PDBReference class, so this code is not duplicated. Updated all testing. There are now tests that evaluate the functions used to initialize pools and that are mapped into the pools. Tests evaluate the ability to produce the desired files and counts as output but do little to evaluate the function of each method, which would be a good improvement in the future (at the moment things seem to be working well).