Updated EvolutionaryTraceAlphabet to add the MultiPositionAlphabet class
This Alphabet class is used to expand any provided alphabet to one which considers multiple positions with characters from the same alphabet. For instance if you want to look at pairs of positions in a protein alignment you would generate a MultiPositionAlphabet with a protein alphabet and size 2 so that the original alphabet is expanded to all possible combinations of 2 amino acids.