FrequencyTable class and tests.
Completed the current version of the class and tests for all methods. One idea that may need to be updated in the future is that the get_count_matrix could return a tensor with dimensions defined by the position_size and alphabet as opposed to just being two dimensional with 1 dimension for positions and another for the alphabet.