Working on Trace class and testing.
Added a function to perform the characterization of the groups at each rank. This is single processed, I would like to make it multiprocessed (tree traversal should be optimized for that though). Current tests on the smaller alignment show that this is working as expected at the moment. I have also tested the initialization function.