Completed multiprocessed version of the Trace.trace function.
This required signfiicant re-writing of the tests and underlying functions. The trace function now creates two pools of workers. The first one only computes the group scores for unique nodes in the tree. Once these are completed and stored in a sharable structure, a second pool is opened that computes the ranks. All tests including a comparison to the WETC identity ranks computation pass successfully at this point.