Updating FrequencyTable so that it can actually hold frequency information not just counts
Added a new attribute to the class called frequencies which is a bool that tracks whether this instance has calculated its frequencies yet. I have updated many functions such that the construction of the __position_table now includes a 'count' field at the pos->char level. I have also updated the __add__ function such that it deletes any computed frequencies in the merged table so that there are not issues with stale frequencies. I updated the tests to reflect the changes in the data structure and so far everything passes (new tests should be added for __add__ to test that the frequency deletion occurs correctly).