Updating FrequencyTable and its testing
In the FrequencyTable class frequency functions were shifted to the end of the class. In addition increment_count was updated so that it changes the frequencies attribute so that one cannot use stale frequencies if new data is added after frequencies are computed. The overwritten __add__ function has been updated to exhaustively check for 'frequency' entries in the __position_table and delete them in any merged table. This is more time/resource intensive but it saves from having to update the table every time increment_count is called and the values become stale. It would also be possible to split the __position_table into two sections one of which can easily be discarded. It might also be good to consider making this class immutable, however this would require more record keeping for other classes which use the FrequencyTable class.