Attempting to make FrequencyTable safe to use in synchronized environments
This is the first time I am attempting this so there are probably better ways but I think this should work. I have added an internal RecursiveLock to all FrequencyTable instances. This is used to wrap every function in an acquire release clause, since the lock is recursive this should support the same process requesting calls from one function to other functions (increment_count) does this for instance. So far all tests still work fine with this new addition.