Updating FrequencyTable and testing
I have made increment_count a private method for FrequencyTable and exposed it through a new function called characterize_sequence. This takes a SeqRecord and and adds all of the positions in the sequence and the characters at those positions. To manage this I have also made the initialization more rigid. It now requires the sequence length of the sequences to be characterized upon initialization. It also allows me to make the tracking of the depth/coverage of the frequency table much easier, since I assume that all sequences which get characterized are unique, I only increase the depth when a characterize_sequence is called. This simplifies __add__ and compute_frequencies.