Updated PositionalScorer changing how rank scores are computed, adding an attribute, and adding several more metrics.
The new attribute to PositionalScorer is the metric type. A metric can either be integer valued ('integer') or real valued ('real_valued'). If a metric is neither of these (as determined by module level sets (integer_metrics and real_valued_metrics) an error is raised during initialization.
Ranks are now computed for either an integer valued metric or a real valued metric (since they tend to be combined in the same way). Because of this rank_identity_score has been renamed rank_integer_score and rank_plain_entropy_score has been renamed rank_real_valued_score.
New methods have been added for pair position scoring, namely: mutual information, normalized mutual information, and average product corrected mutual information (MIp). This includes a new method which computes and returns all intermediate values for the mutual information score (i, j, Hi, Hj, Hij, and mutual information). Since MIp is computed as a modification over the entire MI matrix it has been added as a modifier to the end of the score_group function. The score group function has also been modified to only fill in the upper or lower triangle of a matrix for pair position metrics (i.e. the diagonal is explicitly excluded). normalized MI and MIp have explicit checks for divide by zero errors and raise excpetions if this would occur and the MI was not 0 to begin with.