Updated FrequencyTable and SeqAlignemnt characterize positions
I now cast lil_matrix to a csc_matrix at the end of alignment characterization. This is achieved through a new method (still needs documentation). The behavior is carried out at the end of the SeqAlignment method. I also updated how get_count_array behaves to guarantee proper behavior, changes had to do with order of operations to avoid array type being changed in a way that changes the behavior. I have also updated a few of the tests to increase the speed.