Updated how assign_group_rank works in PhylogeneticTree.
Now the method accepts a variable ranks which specifies a list/set of ranks to analyze. By default (e.g. None) all ranks are analyzed. In addition to allowing only specified ranks to be analyzed this also tracks new information in the assignments dictionary, i.e. the descendants of that node from the most recently processed rank. This was done in an effort to streamline processing during the trace (now each node knows exactly which nodes, that have hopefully already been processed, are components of the desired node), especially now that it will be possible for ranks to be skipped this will be more important. I have updated the assign_group_rank tests to check this new information. I have also added new tests to check what happens when you use a custom ranks list. Finally, to pass these tests it was necessary to update the import_assignments function in ETMIPWrapper.