Updated utils.build_mapping to account for alphabets of size greater than 1.
Did not take into account that if there was an alphabet of size 2 (or greater) like the multiposition alphabet I defined, the gap characters could/would have the same size as the alphabet. I have changed teh behavior of the mapping to make gap characters match the size of the alphabet (This may not be correct in all cases but works correctly for me in this case).