Attempting one other method for characterizing an alignment.
This method appeared slower when I was testing it while producing all the required dictionaries on each call. I have now added the option to pass them in to see if the speed of the old and new characterization methods is comparable. If not I will scrap this and continue.