Updated FrequencyTable and SeqAlignment to finalize new method for characterizing alignment.
The new method is much slower than the previous method for single sequences but scales much better for many sequences (especially when characterizing pairs of sequences). I have updated the documentation for both methods and they are test in the SeqAlignment tests. I do still need to add tests in FrequencyTable for the underlying method.