Updating PhylogeneticTree to Python3, not yet completed.
Cleaned up imports. Changed io for nhx file reading and writing (byte to string change). Traverse_top_down was failing because when there are ties on path length the Clade objects in the tuple are being compared and they cannot be. Fixed this by adding the node name as the second element in the tuple, making the Clade itself the third element (this way if path length is equal hopefully the names won't be, they shouldn't be since there should only be one node with each name). Validate upgma tree tests were failing because using numpy.where is producing a kind of int which DistanceMatrix is interpreting as the wrong type, explicitly casting the positions extracted from the np.where return to int fixed it. Sorted no longer has a cmp keyword it now has key, used the method described here (https://docs.python.org/3/howto/sorting.html) to replace my compare method with a key function in sorted. Traverse bottom us is failing in a similar way as travers_top_down, adding node names the same way I did before, fixes the issues. The get_cluster_spanner method used for converting Agglomerative clustering objects and their returns to nhx trees do not work well in Python3 this still needs to be fixed.