Finishing converting PhylogeneticTree to Python3.
Final change was to make the nhx tree conversion from AgglomerativeClustering work. The difficulty is that one of the key conversion components (AgglomerativeClustering.pooling_func) was deprecated between sklearn 0.19.0 and the current versions of sklearn (0.21.2 in current environment). This component is only used for affinity='euclidean' and linkage='ward' however these are the settings I used most in my previous exploration with this method. My current fix is to hard code the previous default function (np.mean) if the deprecation value (str, 'deprecated') is encountered.