Updated rank score computations and changed nmi computation.
Score ranks now requires a rank variable to be passed in so it can perform normalization correctly. This is needed since rank scoring is no longer performed over a stack of group scores but over a sum of group scores (before the 0th axis of the stack could be used to determine the rank).