Updated .env_template
The template now has many more resources described and better descriptions for each section. Instead of hard coding paths for testing, input, and output they can now be specified through the .env_template. There is also a path and description for the first of the epistasis data files which will be used. Finally, there is a new section including all of the tools which may or may not be needed for running EVCouplings.