Updated PositionalScorer with new scoring method and improvements for other methods.
New scoring metric filtered_average_product_corrected_mutual_information. This recapitulates the ET-MIp implementation of the APC correction to MI by setting all values in the final score matrix to 0 where the MI matrix was less than or equal to 0.0001. To facilitate this and other measures relying on the plain entropy scoring method the group_plain_entropy_score function was updated to work with sparse matrices. This function is still fast and does not cause large memory spikes (scipy.stats.entropy causes large spikes in memory because all matrices are cast to dense). In addition some methods have been cleaned up to remove debugging variables and to make use of variables to improve syntax.