Trace that matches WETC implementations of intET, rvET, and ET-MIp.
This Trace implementation matches WETC scores produced for the options -int, -realval, and -allpairs. Many changes were made to reach this point. rank_scores has been added as an attribute to the class so that the intermediate rank scores can be accessed (like the group scores and characterizations can be reached through the unique_nodes attribute). Characterization has been updated to produce the mappings required by FrequencyTable characterization methods before the processing pool is created. The processing queue for characterization has also been split into two queues one for the lowest (closest to terminal) rank, and another for all higher rank nodes. This has been implemented and test for the case where an incomplete set of ranks is specified (i.e. not the whole tree is used for tracing). Rank scoring has been updated so that 1 is not added to the final ranks for all scoring metrics (mutual_information based scores should not have 1 added since they're meaning is reversed from integer and entropy based methods, i.e. one is the best these scores could achieve instead of the worst). Methods have been added to both the characterization and scoring sections to serialize data (enabling the low_memory option for the class). There are check, save, and load functions for FrequencyTable objects and for numpy.ndarrays. Characterization has the added options now to write out the sub alignments and frequency tables for each node characterized during the trace. To achieve better concurrency behavior in the characterization process, the time it takes to perform the first characterization is recorded such that any time blocking needs to be performed while waiting for descendants to be characterized a fraction of that time can be set as the sleep time as opposed to choosing an arbitrary value or trying to pull the desired value continually. All multiprocessing pools that use queues (all of the ones used in Trace) have been updated to watch for sentinel values ('STOP') instead of checking for empty queues.