Finalizing and testing Trace.
A function has been added to compute the trace and ranks outside of the trace() function. All Trace.trace tests pass. The comparisons to intET and rvET pass. Comparison to ET-MIp fails due to slight differences in the coverage scores which I think are acceptable at the moment (and possible correct, at the moment less than 1 percent of scores for both example structure have a difference of more than 1E-3 which is the precision of ET-MIp).