Updated ETMIPWrapper to correct unexpected behavior and updated test_PhylogeneticTree to work with updated ETMIPWrapper.
To increase speed all calls to ETMIPWrapper in test_PhylogeneticTree use the method 'intET' since this is the fastest and since distance calculation, tree construction, and rank/group assignments are the same for the different methods. Commented out the UPGMA validation of the WETC tree and the 'et' tree building method tree since they are not UPGMA trees and are not expected to validate. Left it in there in case anyone ever wonders about it and wants to run the validation again, or thinks up an improved validation.