Changing imports in test_Base back to relative paths (as opposed to absolute/project paths). Also fixed sorted (cmp to key change).
This allows testing from the command line to work properly without installing the package. The sorted update uses the python documentations suggestion for going from a cmp function to a key function (i.e. wrap the cmp function in a function that returns a class which performs proper evaluation using the cmp method).