Updated PhylogeneticTree agglomerative clustering construction method as well as the testing for rename internal nodes.
The agglomerative clustering method now handles the case where it does not receive a cache dir value. In that case it now creates a file in the current working directory and deletes it after the tree has been loaded. The tests have been updated so that they do not use the construct_tree method but rather each specific contruction method. This also requires the distance matrix and size of the tree to be set manually.