Updated ContactScorer._map_predictions_to_pdb and its tests.
The method in the class has been updated such that it now correctly identifies all mappable positions from the sequence and subsets the predictions array to that subset. It then identifies all mappable positions in the structure and subsets the distance array to that subset. It then finds all pairs in the specified sequence separation category which are mappable and subsets both of the restricted prediction and distance arrays to those pairs. The first test for this passes, the second still needs to be updated.