Updating charcterization, init, and characterize_rank_groups to use tqdm progress bar instead of using print statements.
This significantly reduced the complexity of the characterization function. Now that two queues are not being coordinated simple control flow can be used to determine whether an inner or component node is being processed, which reduces code duplication. An important note is that since the sleep_type variable may be updated during processing (it is protected by the same lock as the sharable dictionary), it is necessary to declare that the global variable is being accessed (unlike in the other methods where they can just be called on). I have also updated the documentation for all methods and functions updated in the last few commits and removed all timing print statements.