Added several new scoring metrics based on "Essential guidelines for computational method benchmarking" by Weber et al.
These include the AUPRC and the tpr/fdr curve. I have implemented area under the curve calculations for both, though technically the tpr/fdr curve is not generally measured in this way. I have updated the evaluate_predictions method to use these in places of the precision and recall scores though I have left F1 computation in the evaluation.