Updated PDBReference to more stringently filter out HETATM entries.
I ran into a case where a HETATM had a standard residue name so it passed my previous filter, but caused an error further down the line. I have added a new check to see if the residue id has an unexpected first entry (usually it is ' ') which starts with 'H_' which all the HETATM residues in the structure I was working with had. Currently the PDBReference test still succeeds.