Updated julia GaussDCA script for DCAWrapper.
The script needed to be updated for the current version of julia so that it called on Pkg correctly to add the package and as well as pathing and git to update it so that it could be properly installed. The calls to GaussDCA and writing the results to file still work properly.