Created a wrapper for calling EVCouplings pipelines in my experiments.
This wrapper is initialzed with an alignment and a protocol name (expecting 'standard' or 'mean_field'). The wrapper is set up to create a set of configurations which passes through an unedited (by using the most permissive settings in the 'align' stage of the evcouplings pipeline) alignment file to the couplings computation (allowing for direct comparison with the other methods which are also fed this alignment). The wrapper also calls the pipeline to execute the specified configurations and imports the scores in the expected format.