Added documentation to the EVCouplingsWrapper as well as updating the delete_files option.
Each method is now commented to explain what the method does and what input it expects. The delete files option for calculate scores has been updated to remove the directories and other output from the pipeline, keeping only data create by the wrapper itself. The large data set cell has also been updated to remove this data which does not need to be stored for the purposes of the experiments in this project.