Updated ContactScorer.
Now all figures previously generated in the .eps format are generated in the .png format. In addition an option 'plots' has been added to evaluate_predictor and evaluate_predictions which allows the user to specify whether to generate plots while computing statistics for a predictor or predictions or whether to just return the statistical scores.