Updated ContactScorer predictor and prediction evaluation.
The evaluations have been changed so verbosity=1 outputs not just AUROC but also AUPRC and AUTPRFDRC. Verbosity=2 outputs not only F1 score but also precision and recall scores at specific top prediction thresholds.Corrections were made such that the AUPRC as correctly calculated from the precision and recall data generated to create the curve. AUTPRFDRC value calculation was corrected as well so that the area reported is the value below (not above) the curve (This value is qualitative as this is not a standradized analysis but rather a curve that is generally inspected by eye, according to papers read up to this point).