Working on ContactScorer and its tests.
Added tests for AUPRC and AUTPRFDRC as well as their plotting functions. The AUTPRFDRC test is not complete and should be returned to later. Updated many of the prior tests to work with the new test proteins. Working on disecting the differences between the original SCW Z-Scoring implementation from PyETViewer and my update for Python3 (and algorithmic updates).