Created a new notebook to look at old alignments.
This is essentially a carbon copy of the Small Test Set Comparison notebook but because it will not use the DataSetGenerator to create the data set and will write out to a new location I have updated those cells. In addition the alignments have a different naming convention for the query sequence which I have captured in the cells where the data is actually generated.