Updated documentation for DataSetGenerator
In addition to updating the documentation I also changed the alignment characterization method to work with decimals for the min and max identity as well as changing the naming of num_threads variables to processes which is more accurate. At the moment there are errors which I have not encountered before which occur during parsing of query sequences that are being read from Swiss/Uniprot, currently troubleshooting these.