Updating Trace and its testing.
The update includes the removal of the infrastructure to support a heuristic scoring method which was in development in the PositionalScorer class but which has also been removed for versioning purposes. This infrastructure and the scoring metric can be added in later in future development if they show promise. In addition I have removed testing for those aspects of the Trace class and have uncommented the remaining tests, which pass at this time.