First attempt to implement group score maximization.
This code is meant to ensure that the covariation score does not get smaller as the tree is traversed. This is done to maintain the property of ET observed in the single position case, where the score always tends towards its optimum as the tree traversal progresses. The scoring metrics used for covariation do not preserve this property so I am attempting to enforce it.