Created a parent class Predictor.py for all predictors.
This was done to enforce a minimum set of attributes and to create a base initializer. It also requires the implementation of a calculate_scores method to homogenize how the different predictors are called. I wrote tests for this parent class which are all currently passing (just that values are initialized properly and that exceptions are raised for unimplemented methods).