Started updating EVCouplingsWrapper to make use of the Predictor parent class.
The EVCouplingsWrapper now has additional attributes and a different init method (leaning heavily on the Predictor super()). Calculate scores has been updated so that it also computes rankings and coverage values for the predictions since I have made that a part of the base set of attributes (I did this partially so it would not accidentally be done incorrectly later in ContactScorer.py or another context). I have created a set of tests for the EVCouplingsWrapper.py which include test for init, import_covariance_score, and calculate_scores. Testing has not yet been completed!