Completed updating ETMIPWrapper.py
All tests have been written and all but the test for importing intermediate scoring values have successfully passed (that tests is incredibly intensive and often fails due simply to memory constraints, it is likely that the method should be re-thought and likely removed). The test for caclulate scores is lenient partially because the answers written to file are rounded so the absolute scores can only be compared two three decimal places while only the trend of the ranks can be compared directly.