Removed the Maximize option and created a trace visualization method.
The maximize function has been removed from Trace.trace and has now been removed from the EvolutionaryTrace implementation and testing. I have also created a function which generates a visualization of the trace (for specified ranks, or all ranks) such that the figures keep their size and progressively fill in groups from one rank at a time so that the break down of the alignment over the tree traversal can be easily visualized for a given position/set of positions. Tests for the EvolutionaryTrace.py and its new visualization method pass current tests.