Tests for the MatchMismatchTable.py have been updated and passed.
For the class and for the tests the characterize matches and mismatches methods have been commented out as I do not think this is the best way of performing this characterization. I have also commented out the subset method. I plan to remove these once I have successfully implemented characterization in the Trace algorithm.