First full draft of characterization of matches and mismatches.
The original characterize_rank_groups is now a switch statement for either the standard characterization of columns by nucleic/amino acid content or the characterization of all column pairs for matches and mismatches. The new characterization required a new init function for multiprocessing pools and a function which can be mapped to. This first draft relies on functions which mirror the previous implementation quite closely. An alternative implementation was started before this which is still present but currently commented out, it seemed like it might be faster but would require more troubleshooting because it did not handle the large memory requirements addressed by the current implementation.