Updated FrequencyTable.py to change the alphabet checks and to update functions like get_chars.
Alphabet checks now require only that the reverse mapping be the same size as the alphabet and that the mapping be the same size or larger (to include gap characters) than the alphabet. It also requires that the index of the first element in the mapping return the proper character from the reverse mapping. get_chars has been updated to make use of the new form the reverse mapping (array instead of dict). The resulting function is slightly faster than the list comprehension using a dictionary used previously.