Changing FrequencyTable.py increment_count and characterize_alignment.
The increment_count method now includes a check to make sure the table has not already been finalized before it continues, if so an AttributeError is raised (I chose this exception because the position_table attribute is no longer the correct type for updating). I have also updated characterize_alignment to work differently in two ways. First, instead of adding the counts of each character at a specific instance I am adding each instance separately because it is faster (this may cause a memory error, I won't know until I test larger examples). The second change is that the single_to_pair mapping is no longer a dictionary but a matrix where the shape of the matrix is a square with the length of the single letter alphabet and the contents are the numerical values for the pair alphabet mapping, this allows a full column worth of characters to be translated at once simply by providing each column as indices to the matrix. This will require updating inputs in other code (check Trace and EvolutionaryTrace).