Updating MatchMismatchTable.py and its testing.
The update to the class includes enforcing access to only the upper triangle for get_status_and_character, an error is now raised if seq_ind1 is greater than or equal to seq_ind2. Testing has been updated to significantly reduce the time spent on assessing get_status_and_character, instead of checking all sequence comparisons for all positions in the alignment (for single or pair) errors are enforced for lower triangle access and one random comparison from the upper triangle is performed for each of the alignment positions. Currently all tests are passing.