Updated Trace.py to properly passing mappings to the different characterization methods.
Also had to correct how the trace function and its supporting functions access the FrequencyTables so that the right ones are accessed depending on if match_mismatch has been specified or not. Tests have been updated and reduced to reduce testing time and to properly test the match mismatch cases.