Updated EvolutionaryTrace.py and its testing.
This update ensures that the match/mismatch characterization and scoring are correctly called in the trace. It also updates how results (variation column) are written to file (using the FrequencyTable) so that it works for both the standard and match/mismatch case. Tests have also been updated to ensure that the match/mismatch case is tested properly for each of the EvolutionaryTrace methods.