Did my best to shift the parsing of external sequence identifiers for chains in a PDB structure and the corresponding sequences from the DataSetGenerator.py class to the PDBReference.py class.
This required a get_sequence method as well as a series of accession and sequence parsing methods for Swiss/Uniprot and GenBank. Simple tests were written and are passing for these behaviors. The recommended biopython procedure for parsing Swiss/Uniprot sequences from their references is no longer working correctly, so I wrote a very short regular expression method to retrieve amino acid sequences, I am not sure how robust it is but it did work for the test cases I wrote.