Updated implementation of Trace.py and its tests to improve the speed and efficiency of match/mismatch characterization.
The new characterization works by realizing any inner node's match and mismatch FrequencyTables can be generated by adding the FrequencyTables for its children (direct descendants) which covers part of the new upper triangle represented by the parent node and then adding data from the rectangle(s) which fill in the rest of that upper triangle. This required some updates to a number of the characterization methods and resulted in the standard and match/mismatch characterization protocols looking even more similar. In most cases this seems to provide a significant speed up.