New methods for characterizing matches and mismatches given a set of sequence identifier indices (corresponding to the overall alignment).
These new methods allow for the return of characterization on single positions or multi-positions. A single position characterization returns exactly the single letter alphabet for the sequences specified in each comparison and the match (1) or mismatch (-1) of the characters observed there. For a pair or larger position the characters returned are the composite of the single position characterizations while the match/mismatch return is the sum of the single letter match/mismatch status. Additional methods have been added to simplify calling this for sub-alignments, one returns the characterization of a triangle within the upper triangle (specified by a single set of sequence identifier indices, i.e. one sub-alignment), while another returns the characterization of a rectangle within the upper triangle (specified by two sets of sequence identifiers, i.e. two sub-alignments). All new methods have been documented and tested.