Changed Trace.py match/mismatch characterization again as well as the pooling method testing.
This new implementation goes back to the current implementation in the master branch, and removes the waiting for descendants from the multiprocessing function. These updates are performed in the main thread after all sub-problems have returned. The reason this is being done is that I hope things will finish more quickly if large chunks of the problem are not waiting on some particularly large sub-problem in the middle. I am also assuming that updating the data using completed sub-problems in the main thread will be relatively fast. This still needs to be tested.