Reworking utils.py testing.
So far I have started writing tests for build mapping trying to fit TTD principles. The first change this has led to is that alphabets which are strings or lists are coerced to lists instead of strings. At the moment there are still tests failing for skip letters which overlap the alphabet and several two letter alphabet tests. These are next on the todo list.