Reworking utils.py testing.
I have completed writing tests for build mapping. I now cover not only the single letter string alphabets, but one and two letter list alphabets, as well as one and three letter alphabets using the Bio.Alphabet class. One refactoring was needed to ensure that skip_letters were removed before the mapping and alphabet size were determined so that the size and mapping would not be artificially high if there was a skip letter from the alphabet.