Reworking EvolutionaryTraceAlphabet.py testing.
I have written tests for the static single letter alphabets (DNA and Protein) to ensure that they contain the expected characters. I have also written tests for multi letter alphabets starting from single and multi letter alphabets, with sizes 1, 2, and 4, and have tested using strings, lists, and alphabet objects as the input. I have updated the documentation to explain that size is really a multiplier of the input alphabet, so the new size will be the size of the previous alphabet characters, times the new size specified.