Started testing PositionalScorer.py
Wrote tests for the top level variables as well as the __init__ function. One additional check was added to init to ensure that position sizes > 2 are not used at this time. Additional metrics which I plan to add during this update were also added to the top level variables for tracking available metrics.