Testing PositionalScorer.py
Completed tests for group_plain_entropy score as well as for mutual information scoring functions including: mutual_information_computation, group_mutual_information_score, group_normalized_mutual_information_score, average_product_correction, and filtered_average_product_correction. A check was added to group_plain_entropy_score to ensure that the dimensions passed matched the frequency table provided, and a correction was made to one of the error messages. Checks were added to the mutual_information_computation to ensure that the frequency table provided matched the dimensions provided and which enforce a dimension size of 2 for this function. A check was added to average_product_correction to ensure that the provided matrix is not None and that only the upper triangle (excluding the diagonal) is initialized. filtered_average_product_correction was significantly reduced, it now calls average_product_correction internally and updates any positions which do not pass the threshold. The threshold is now exposed to the user as a variable (which defaults to the previously used value).