Testing and Updating PositionalScorer.py
Moved the group_match_count and group_mismatch_count functions to the real valued scoring function group. I am doing this because the scores are not meant to be interpreted in a binary way, their magnitude has meaning. I have updated the corresponding tests (for __init__). I have also corrected a test which was not working properly: test_group_match_mismatch_diversity_angle_mismtach_zeros, the expectation was set incorrectly but is now correct.